+++ test/clone.go
@@ -1,5 +1,7 @@
 package test
 
+import "time"
+
 func CloneBool(datum *bool) *bool {
 	if datum == nil {
 		return nil
@@ -40,5 +42,13 @@ func CloneStringArray(datum *[]string) *[]string {
 	for index, value := range *datum {
 		clone[index] = value
 	}
+	return &clone
+}
+
+func CloneTime(datum *time.Time) *time.Time {
+	if datum == nil {
+		return nil
+	}
+	clone := *datum
 	return &clone
 }